Beginning to refactor structs to storage folder
