+++ main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/team7mysupermon/mySuperMon_Middleware/storage"
 	"io/ioutil"
 	"net/http"
 	"strings"
@@ -12,62 +13,27 @@ import (
 	"github.com/team7mysupermon/mySuperMon_Middleware/monitoring"
 )
 
-/*
-Struct to get all information from the bearer token returned by MySuperMon upon login
-It must match the json object that we get from MySuperMon
-*/
-type Token struct {
-	AccessToken string `json:"access_token"`
-	Type        string `json:"token_type"`
-	ExpiresIn   int    `json:"expires_in"`
-	Scope       string `json:"scope"`
-	Jti         string `json:"jti"`
-}
-
-/*
-Struct for the calls that the user makes with a Usecase and a ApplicationIdentifier.
-The information is gathered from the URL and added to the struct.
-Therefore, the end of the struct, the uri:xxx part, must match the definitions in the main method.
-*/
-type StartAndStopCommand struct {
-	Usecase               string `uri:"Usecase" binding:"required"`
-	ApplicationIdentifier string `uri:"Appiden" binding:"required"`
-}
-
-/*
-Struct for the calls that the user makes with a Username and a Password.
-The information is gathered from the URL and added to the struct.
-Therefore, the end of the struct, the uri:xxx part, must match the definitions in the main method.
-*/
-
-type LoginCommand struct {
-	Username string `uri:"Username" binding:"required"`
-	Password string `uri:"Password" binding:"required"`
-}
-
-
 var (
 	// The authentication token needed to be able to get the access token when logging in
 	authToken = "Basic cGVyZm9ybWFuY2VEYXNoYm9hcmRDbGllbnRJZDpsamtuc3F5OXRwNjEyMw=="
 
 	/*
-	Instantiated when a user calls the login API call.
-	Contains the authentication token
+		Instantiated when a user calls the login API call.
+		Contains the authentication token
 	*/
-	Tokenresponse Token
+	Tokenresponse storage.Token
 
 	/*
-	Closes the goroutine that scrapes the recording.
-	The goroutine is started when the user starts the recording
+		Closes the goroutine that scrapes the recording.
+		The goroutine is started when the user starts the recording
 	*/
 	quit = make(chan bool)
-) 
-
+)
 
 func main() {
 	go monitoring.Monitor()
 
-	router := gin.Default()	
+	router := gin.Default()
 
 	// The API calls
 	router.GET("/Login/:Username/:Password", getAuthToken)
@@ -75,13 +41,16 @@ func main() {
 	router.GET("/Stop/:Usecase/:Appiden", stopRecording)
 
 	// Starts the program
-	router.Run(":8999")
+	err := router.Run(":8999")
+	if err != nil {
+		return
+	}
 }
 
 func startRecording(c *gin.Context) {
 	// Creates the command structure by taking information from the URL call
 	// TODO: Handle errors
-	var command StartAndStopCommand
+	var command storage.StartAndStopCommand
 	if err := c.ShouldBindUri(&command); err != nil {
 		c.JSON(400, gin.H{"msg": err})
 		return
@@ -96,7 +65,7 @@ func startRecording(c *gin.Context) {
 func stopRecording(c *gin.Context) {
 	// Creates the command structure by taking information from the URL call
 	// TODO: Handle errors
-	var command StartAndStopCommand
+	var command storage.StartAndStopCommand
 	if err := c.ShouldBindUri(&command); err != nil {
 		c.JSON(400, gin.H{"msg": err})
 		return
@@ -114,7 +83,7 @@ func getAuthToken(c *gin.Context) {
 
 	// Creates the command structure by taking information from the URL call
 	// TODO: Handle errors
-	var command LoginCommand
+	var command storage.LoginCommand
 	if err := c.ShouldBindUri(&command); err != nil {
 		c.JSON(400, gin.H{"msg": err})
 		return
@@ -148,7 +117,10 @@ func getAuthToken(c *gin.Context) {
 		fmt.Println(err)
 		return
 	}
-	json.Unmarshal(body, &Tokenresponse)
+	err = json.Unmarshal(body, &Tokenresponse)
+	if err != nil {
+		return
+	}
 
 	fmt.Println("******************************************** Auth Token ********************************************")
 	fmt.Printf("%s : %s\n", Tokenresponse.Type, Tokenresponse.AccessToken)
@@ -183,7 +155,6 @@ func Operation(usecase string, action string, applicationIdentifier string) *htt
 		}
 	}
 	defer res.Body.Close()
-	
 
 	// TODO: Handle errors
 	body, err := ioutil.ReadAll(res.Body)
@@ -208,7 +179,7 @@ func Operation(usecase string, action string, applicationIdentifier string) *htt
 The function that is called when the user starts the recording
 Will every 5 seconds do the run operation, which returns some information about the current recording
 */
-func scrapeWithInterval(command StartAndStopCommand) {
+func scrapeWithInterval(command storage.StartAndStopCommand) {
 	for {
 		select {
 		case <-quit:
@@ -227,4 +198,3 @@ func generateUserInfo(username string, password string) string {
 
 	return userInfo
 }
-
